Move conditional decorator to testing_support
